Update Product and DataBase with his children: change price type, because in AddProductServlet it was long -> type should be long (stupid student).

+++ src/main/java/ru/akirakozov/sd/refactoring/databse/DataBase.java
@@ -6,8 +6,8 @@ import java.util.Optional;
 public interface DataBase {
     public void                 addProduct(Product product);
     public List<Product>        selectAllProducts();
-    public Integer              getProductsCount();
-    public Integer              getProductPricesSum();
+    public int                  getProductsCount();
+    public long                 getProductPricesSum();
     public Optional<Product>    getProductWithMinPrice();
     public Optional<Product>    getProductWithMaxPrice();
     public void                 dropTable();
